feat: add foam container to machine

+++ src/containers/cacao_container.rs
@@ -69,7 +69,7 @@ impl CacaoContainer {
     }
 
     fn check_capacity(&self) -> bool {
-        let min_capacity = (N as f32) * (0.2 as f32);
+        let min_capacity = (N as f32) * (0.2_f32);
         self.capacity as f32 <= min_capacity
     }
 }
